demos/the-predictive-lambda/requirements.txt
-e .
aws-cdk-lib==2.50.0
path==13.1.0
aws-cdk.awsapigatewayv2-alpha==2.50.0a0
aws-cdk.awsapigatewayv2-integrations-alpha==2.50.0a0

